requirements.txt
# Requirements for the project
fastapi==0.112.0
uvicorn[standard]==0.30.5
psycopg[binary,pool]==3.2.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
bcrypt
python-dotenv
stripe